// src/store/categories/interface.ts
import {ImageDetails} from "../../types.ts";


export type ICategory = {
    id: number
    parentId: number
    orderBy: number
    hdThumbnailUrl: string
    thumbnailUrl: string
    imageUrl: string
    originalImageUrl: string
    originalImage: ImageDetails
    name: string
    thumbnail: ImageDetails
    nameTranslated: string
    origin: string
    url: string
    autogeneratedSlug: string
    customSlug: string
    productCount: number
    enabledProductCount: number
    description: string
    descriptionTranslated: string
    enabled: boolean
    productIds: number[]
    seoTitle: string
    seoTitleTranslated: string
    seoDescription: string
    seoDecriptionTranslated: string
    alt: string
}

export interface State {
    categories: ICategory[],
    isLoading: boolean,
    error: string,
}